# perf(escaping): Remove catastrophic backtracking regex for code escape