@SystemKeeper 2025-02-25
Missed the non-greedy quantifier 😓  Added a test for that.